databaseConstruction

Construct MS2 spectra database according to mzXML data and compound information table (csv format).

databaseConstruction(path = ".", version = "0.0.1",
  metabolite.info.name = "metabolite.info.csv",
  source = "Michael Snyder Lab",
  link = "http://snyderlab.stanford.edu/", creater = "Xiaotao Shen",
  email = "shenxt1990@163.com", rt = TRUE, mz.tol = 15,
  rt.tol = 30, threads = 3)

Arguments

path
Work directory.

version
The version of you database. Default is 0.0.1.

metabolite.info.name
The metabolite information table name, it must be csv format. The demo data can be got from the `demoData` package. Please see https://jaspershen.github.io/metID/articles/metID.html

source
The source of your database.

link
Website link of the source.

creater
Creater name. For example, Xiaotao Shen.

email
email address.

rt
Do the metabolites have RT information or not?. If not, set it as FALSE.

mz.tol
m/z tolerance for the match between metabolites and precursor m/z of MS2 spectra.

rt.tol
RT tolerance for the match between metabolites and precursor m/z of MS2 spectra.

threads
The number of threads

Value

A databaseClass object.

See also