Conflicts between the metid and other packages

This function lists all the conflicts between packages in the metid and other packages that you have loaded.

metid_conflicts()

Value

metid_conflicts

Details

There are four conflicts that are deliberately ignored: intersect, union, setequal, and setdiff from dplyr. These functions make the base equivalents generic, so shouldn't negatively affect any existing code.

Examples

metid_conflicts()
#> ── Conflicts ────────────────────────────────────────────── metid_conflicts() ──
#> ✖ stats4::AIC()               masks stats::AIC()
#> ✖ stats4::BIC()               masks stats::BIC()
#> ✖ BiocGenerics::IQR()         masks stats::IQR()
#> ✖ BiocGenerics::Position()    masks ggplot2::Position(), base::Position()
#> ✖ S4Vectors::aggregate()      masks stats::aggregate()
#> ✖ MSnbase::all.equal()        masks base::all.equal()
#> ✖ massdataset::apply()        masks base::apply()
#> ✖ ProtGenerics::as.list()     masks S4Vectors::as.list(), BiocGenerics::as.list(), base::as.list()
#> ✖ methods::body<-()           masks base::body<-()
#> ✖ BiocGenerics::boxplot()     masks graphics::boxplot()
#> ✖ stats4::coef()              masks stats::coef()
#> ✖ BiocGenerics::colMeans()    masks massdataset::colMeans(), base::colMeans()
#> ✖ BiocGenerics::colSums()     masks massdataset::colSums(), base::colSums()
#> ✖ MSnbase::colnames<-()       masks S4Vectors::colnames<-(), BiocGenerics::colnames<-(), base::colnames<-()
#> ✖ MSnbase::combine()          masks Biobase::combine(), BiocGenerics::combine(), dplyr::combine()
#> ✖ S4Vectors::complete.cases() masks stats::complete.cases()
#> ✖ stats4::confint()           masks stats::confint()
#> ✖ S4Vectors::cor()            masks stats::cor()
#> ✖ S4Vectors::cov()            masks stats::cov()
#> ✖ BiocGenerics::density()     masks stats::density()
#> ✖ S4Vectors::end()            masks BiocGenerics::end(), stats::end()
#> ✖ massdataset::filter()       masks dplyr::filter(), metid::filter(), stats::filter()
#> ✖ S4Vectors::first()          masks dplyr::first()
#> ✖ S4Vectors::head()           masks utils::head()
#> ✖ MSnbase::image()            masks BiocGenerics::image(), graphics::image()
#> ✖ S4Vectors::intersect()      masks BiocGenerics::intersect(), massdataset::intersect(), base::intersect()
#> ✖ methods::kronecker()        masks base::kronecker()
#> ✖ dplyr::lag()                masks stats::lag()
#> ✖ stats4::logLik()            masks stats::logLik()
#> ✖ BiocGenerics::mad()         masks stats::mad()
#> ✖ S4Vectors::na.exclude()     masks stats::na.exclude()
#> ✖ S4Vectors::na.omit()        masks stats::na.omit()
#> ✖ stats4::nobs()              masks stats::nobs()
#> ✖ MSnbase::plot()             masks stats4::plot(), graphics::plot(), base::plot()
#> ✖ stats4::profile()           masks stats::profile()
#> ✖ Rcpp::prompt()              masks utils::prompt()
#> ✖ BiocGenerics::relist()      masks utils::relist()
#> ✖ S4Vectors::rename()         masks massdataset::rename(), dplyr::rename()
#> ✖ BiocGenerics::residuals()   masks stats::residuals()
#> ✖ BiocGenerics::rowMeans()    masks massdataset::rowMeans(), base::rowMeans()
#> ✖ BiocGenerics::rowSums()     masks massdataset::rowSums(), base::rowSums()
#> ✖ MSnbase::scale()            masks base::scale()
#> ✖ S4Vectors::sd()             masks BiocGenerics::sd(), stats::sd()
#> ✖ MSnbase::smooth()           masks ProtGenerics::smooth(), stats::smooth()
#> ✖ MSnbase::split()            masks S4Vectors::split(), base::split()
#> ✖ S4Vectors::stack()          masks utils::stack()
#> ✖ S4Vectors::start()          masks BiocGenerics::start(), stats::start()
#> ✖ MSnbase::t()                masks S4Vectors::t(), BiocGenerics::t(), base::t()
#> ✖ S4Vectors::tail()           masks utils::tail()
#> ✖ MSnbase::trimws()           masks base::trimws()
#> ✖ stats4::update()            masks stats::update()
#> ✖ S4Vectors::var()            masks BiocGenerics::var(), stats::var()
#> ✖ stats4::vcov()              masks stats::vcov()
#> ✖ BiocGenerics::weights()     masks stats::weights()
#> ✖ S4Vectors::window()         masks stats::window()
#> ✖ S4Vectors::xtabs()          masks BiocGenerics::xtabs(), stats::xtabs()